+++ grid2op/Parameters.py
@@ -118,7 +118,7 @@ class Parameters:
         # this is expressed in relative value of the thermal limits
         # for example setting "HARD_OVERFLOW_THRESHOLD = 2" is equivalent, if a powerline has a thermal limit of
         # 243 A, to disconnect it instantly if it has a powerflow higher than 2 * 243 = 486 A
-        self.HARD_OVERFLOW_THRESHOLD = dt_int(2)
+        self.HARD_OVERFLOW_THRESHOLD = dt_float(2.)
 
         # are the powerflow performed by the environment in DC mode (dc powerflow) or AC (ac powerflow)
         self.ENV_DC = False
@@ -310,3 +310,110 @@ class Parameters:
         """
         res = Parameters(json_path)
         return res
+
+    def check_valid(self):
+        """
+
+        check the parameter is valid (ie it checks that all the values are of correct types and within the
+        correct range.
+
+        Raises
+        -------
+        An exception if the parameter is not valid
+        """
+        try:
+            if not isinstance(self.NO_OVERFLOW_DISCONNECTION, (bool, dt_bool)):
+                raise RuntimeError("NO_OVERFLOW_DISCONNECTION should be a boolean")
+            self.NO_OVERFLOW_DISCONNECTION = dt_bool(self.NO_OVERFLOW_DISCONNECTION)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert NO_OVERFLOW_DISCONNECTION to bool with error \n:\"{exc_}\"")
+
+        try:
+            self.NB_TIMESTEP_OVERFLOW_ALLOWED = int(self.NB_TIMESTEP_OVERFLOW_ALLOWED)  # to raise if numpy array
+            self.NB_TIMESTEP_OVERFLOW_ALLOWED = dt_int(self.NB_TIMESTEP_OVERFLOW_ALLOWED)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert NB_TIMESTEP_OVERFLOW_ALLOWED to int with error \n:\"{exc_}\"")
+
+        if self.NB_TIMESTEP_OVERFLOW_ALLOWED < 0:
+            raise RuntimeError("NB_TIMESTEP_OVERFLOW_ALLOWED < 0., this should be >= 0.")
+        try:
+            self.NB_TIMESTEP_RECONNECTION = int(self.NB_TIMESTEP_RECONNECTION)  # to raise if numpy array
+            self.NB_TIMESTEP_RECONNECTION = dt_int(self.NB_TIMESTEP_RECONNECTION)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert NB_TIMESTEP_RECONNECTION to int with error \n:\"{exc_}\"")
+        if self.NB_TIMESTEP_RECONNECTION < 0:
+            raise RuntimeError("NB_TIMESTEP_RECONNECTION < 0., this should be >= 0.")
+        try:
+            self.NB_TIMESTEP_COOLDOWN_LINE = int(self.NB_TIMESTEP_COOLDOWN_LINE)
+            self.NB_TIMESTEP_COOLDOWN_LINE = dt_int(self.NB_TIMESTEP_COOLDOWN_LINE)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert NB_TIMESTEP_COOLDOWN_LINE to int with error \n:\"{exc_}\"")
+        if self.NB_TIMESTEP_COOLDOWN_LINE < 0:
+            raise RuntimeError("NB_TIMESTEP_COOLDOWN_LINE < 0., this should be >= 0.")
+        try:
+            self.NB_TIMESTEP_COOLDOWN_SUB = int(self.NB_TIMESTEP_COOLDOWN_SUB)  # to raise if numpy array
+            self.NB_TIMESTEP_COOLDOWN_SUB = dt_int(self.NB_TIMESTEP_COOLDOWN_SUB)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert NB_TIMESTEP_COOLDOWN_SUB to int with error \n:\"{exc_}\"")
+        if self.NB_TIMESTEP_COOLDOWN_SUB < 0:
+            raise RuntimeError("NB_TIMESTEP_COOLDOWN_SUB < 0., this should be >= 0.")
+        try:
+            self.HARD_OVERFLOW_THRESHOLD = float(self.HARD_OVERFLOW_THRESHOLD)  # to raise if numpy array
+            self.HARD_OVERFLOW_THRESHOLD = dt_float(self.HARD_OVERFLOW_THRESHOLD)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert HARD_OVERFLOW_THRESHOLD to float with error \n:\"{exc_}\"")
+        if self.HARD_OVERFLOW_THRESHOLD < 1.:
+            raise RuntimeError("HARD_OVERFLOW_THRESHOLD < 1., this should be >= 1. (use env.set_thermal_limit "
+                               "to modify the thermal limit)")
+        try:
+            if not isinstance(self.ENV_DC, (bool, dt_bool)):
+                raise RuntimeError("NO_OVERFLOW_DISCONNECTION should be a boolean")
+            self.ENV_DC = dt_bool(self.ENV_DC)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert ENV_DC to bool with error \n:\"{exc_}\"")
+        try:
+            self.MAX_SUB_CHANGED = int(self.MAX_SUB_CHANGED)  # to raise if numpy array
+            self.MAX_SUB_CHANGED = dt_int(self.MAX_SUB_CHANGED)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert MAX_SUB_CHANGED to int with error \n:\"{exc_}\"")
+        if self.MAX_SUB_CHANGED < 0:
+            raise RuntimeError("MAX_SUB_CHANGED should be >=0 (or -1 if you want to be able to change every "
+                               "substation at once)")
+        try:
+            self.MAX_LINE_STATUS_CHANGED = int(self.MAX_LINE_STATUS_CHANGED)  # to raise if numpy array
+            self.MAX_LINE_STATUS_CHANGED = dt_int(self.MAX_LINE_STATUS_CHANGED)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert MAX_LINE_STATUS_CHANGED to int with error \n:\"{exc_}\"")
+        if self.MAX_LINE_STATUS_CHANGED < 0:
+            raise RuntimeError("MAX_LINE_STATUS_CHANGED should be >=0 "
+                               "(or -1 if you want to be able to change every powerline at once)")
+        try:
+            if not isinstance(self.IGNORE_MIN_UP_DOWN_TIME, (bool, dt_bool)):
+                raise RuntimeError("IGNORE_MIN_UP_DOWN_TIME should be a boolean")
+            self.IGNORE_MIN_UP_DOWN_TIME = dt_bool(self.IGNORE_MIN_UP_DOWN_TIME)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert IGNORE_MIN_UP_DOWN_TIME to bool with error \n:\"{exc_}\"")
+        try:
+            if not isinstance(self.ALLOW_DISPATCH_GEN_SWITCH_OFF, (bool, dt_bool)):
+                raise RuntimeError("ALLOW_DISPATCH_GEN_SWITCH_OFF should be a boolean")
+            self.ALLOW_DISPATCH_GEN_SWITCH_OFF = dt_bool(self.ALLOW_DISPATCH_GEN_SWITCH_OFF)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert ALLOW_DISPATCH_GEN_SWITCH_OFF to bool with error \n:\"{exc_}\"")
+
+        try:
+            self.INIT_STORAGE_CAPACITY = float(self.INIT_STORAGE_CAPACITY)  # to raise if numpy array
+            self.INIT_STORAGE_CAPACITY = dt_float(self.INIT_STORAGE_CAPACITY)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert INIT_STORAGE_CAPACITY to float with error \n:\"{exc_}\"")
+
+        if self.INIT_STORAGE_CAPACITY < 0.:
+            raise RuntimeError("INIT_STORAGE_CAPACITY < 0., this should be within range [0., 1.]")
+        if self.INIT_STORAGE_CAPACITY > 1.:
+            raise RuntimeError("INIT_STORAGE_CAPACITY > 1., this should be within range [0., 1.]")
+
+        try:
+            if not isinstance(self.ACTIVATE_STORAGE_LOSS, (bool, dt_bool)):
+                raise RuntimeError("ACTIVATE_STORAGE_LOSS should be a boolean")
+            self.ACTIVATE_STORAGE_LOSS = dt_bool(self.ACTIVATE_STORAGE_LOSS)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert ACTIVATE_STORAGE_LOSS to bool with error \n:\"{exc_}\"")