+++ grid2op/data/educ_case14_storage/config.py
@@ -1,4 +1,4 @@
-from grid2op.Action import PowerlineChangeAndDispatchAction
+from grid2op.Action import PowerlineChangeDispatchAndStorageAction
 from grid2op.Reward import L2RPNReward
 from grid2op.Rules import DefaultRules
 from grid2op.Chronics import Multifolder
@@ -7,7 +7,7 @@ from grid2op.Backend import PandaPowerBackend
 
 config = {
     "backend": PandaPowerBackend,
-    "action_class": PowerlineChangeAndDispatchAction,
+    "action_class": PowerlineChangeDispatchAndStorageAction,
     "observation_class": None,
     "reward_class": L2RPNReward,
     "gamerules_class": DefaultRules,